batch/requirements.txt
httpx==0.27.0
beautifulsoup4==4.12.3
lxml==5.2.0
supabase==2.4.0
google-cloud-vision==3.7.2
Pillow==10.3.0
python-dotenv==1.0.1
